Add log file output option
:Release Notes:
This update introduces an option to output logs to a file.

:Detailed Notes:
A new command-line option '-e' is added to specify an output log file
for capturing standard output. The main logic now checks if this option
is provided, redirects 'std::cout' to the specified file, and restores
it afterward. This enhances the tool's usability for debugging and
record keeping.

:Testing Performed:
Local Build & Test : OK
http://jira.lge.com/issue/browse/QWTC-15593

:QA Notes:
N/A

:Issues Addressed:
[WRR-16107] [ACP Reference] Enable Camera Valid TCs

Change-Id: Ibd51f04bf73e8b2349d3670f68de04e7ae3daa5c